enh: add first login info, break tests into files
