+++ profiles/app.py
@@ -1,5 +1,7 @@
 from django.apps import AppConfig
 
-
 class ProfilesConfig(AppConfig):
     name = 'profiles'
+
+    def ready(self):
+        from . import signals
