+++ profiles/tests/test_account.py
@@ -6,7 +6,7 @@ from django.contrib.messages import get_messages, constants
 from profiles.models import Profile, Recommendation, User
 
 
-class SignupViewTests(TestCase):
+class AccountTests(TestCase):
 
     def test_create(self):
         response = self.client.get(reverse('profiles:signup'))
@@ -47,13 +47,22 @@ class SignupViewTests(TestCase):
         self.assertEqual(len(messages), 1)
         self.assertEqual(messages[0].level, constants.ERROR)
 
-
         response = self.client.get(reverse('profiles:signup_confirm'), data={
             'uid': uid,
             'token': token,
         })
         self.assertEqual(response.status_code, HTTPStatus.FOUND)
         self.assertEqual(response.url, reverse('profiles:user'))
+        self.assertTrue('first_login' in self.client.session)
+
+        self.client.session.flush()
+        self.client.logout()
+
+        u = User.objects.get(email='test@test.com')
+        self.client.force_login(u)
+
+        self.assertIsNotNone(u.last_login)
+        self.assertFalse('first_login' in self.client.session)
 
     def test_create_different_session(self):
 
@@ -160,55 +169,3 @@ class SignupViewTests(TestCase):
         self.assertEqual(response.status_code, HTTPStatus.FOUND)
         self.assertEqual(response.url, reverse('profiles:user'))
 
-    def test_profile_delete(self):
-
-        u = User(email='test@test.com')
-        u.is_active = True
-        u.save()
-
-        p = Profile(contact_email='test@test.com', user=u)
-        p.save()
-
-        r = Recommendation(profile=p, reviewer_email='another@test.com')
-        r.save()
-
-        self.client.force_login(u)
-
-        response = self.client.get(reverse('profiles:user_profile_delete'))
-        self.assertEqual(response.status_code, HTTPStatus.OK)
-
-        User.objects.get(id=u.id)
-
-        response = self.client.post(reverse('profiles:user_profile_delete'), data={
-            'confirm': True,
-        })
-        self.assertEqual(response.status_code, HTTPStatus.FOUND)
-        self.assertEqual(response.url, reverse('profiles:user'))
-
-        p.refresh_from_db()
-        self.assertNotEquals(p.deleted_at, None)
-
-    def test_profile_delete(self):
-
-        u = User(email='test@test.com')
-        u.is_active = True
-        u.save()
-
-        p = Profile(contact_email='test@test.com', user=u)
-        p.save()
-
-        self.client.force_login(u)
-
-        response = self.client.get(reverse('profiles:user_profile_delete'))
-        self.assertEqual(response.status_code, HTTPStatus.OK)
-
-        User.objects.get(id=u.id)
-
-        response = self.client.post(reverse('profiles:user_profile_delete'), data={
-            'confirm': True,
-        })
-        self.assertEqual(response.status_code, HTTPStatus.FOUND)
-        self.assertEqual(response.url, reverse('profiles:user'))
-
-        with self.assertRaises(Profile.DoesNotExist):
-            p.refresh_from_db()
